refac: moved settings interfaces & updated imports
