+++ src/interfaces/app/ShowdexShowdownSettings.ts
@@ -1,0 +1,14 @@
+/**
+ * Showdown-specific settings.
+ *
+ * @since 1.1.7
+ */
+export interface ShowdexShowdownSettings {
+  /**
+   * Whether to auto-accept Open Team Sheets when prompted by the server.
+   *
+   * @default false
+   * @since 1.1.7
+   */
+  autoAcceptSheets: boolean;
+}
